Remove badness on branch repeat
Because repeating branches are always finished, or removed because of
repetition inside the same branch, the automatic badness increment is
unneeded.

This badness introduces a case where a thread is merged with another
thread with higher badness, causing the ordering of trace prefixes to
be off. (see test case in Example.tests.js)

By removing the badness, the ordering of trace prefixes is by the time
the branches arrive at a joining instruction. Because this is ordered by
the badness of their respective threads, prefixes are ordened by
badness.