Clear threads before they are put in the oldThreadsList
As the Trace of a Thread is recreated when the Thread is recycled, it
is better to clear the reference to the old Trace when a Thread is put
in the oldThreadsList. This allows the garbage collector to claim the
old Trace.

Additionally, we should not compact these traces, as only the last set
will actually be returned to the caller. Instead, Result now compacts
only the accepting and failing traces at the end of the run.